Implemented all functionality to meet requirements of phase 7: The program should be updated to...
Prompt the user for a difficulty level before they are prompted to guess the number.
The difficulty level should determine how many guesses the user gets. The difficulty levels should be:
Easy - this gives the user eight guesses.
Medium - this gives the user six guesses.
Hard - this gives the user four guesses.